refactor settings code
