+++ plugin.php
@@ -36,7 +36,7 @@ or at 5000 Yonge Street, Suite 1901, Toronto, Ontario, M2N 7E9, Canada.
 */
 
 require_once plugin_dir_path(__FILE__) . 'src/class-activator.php';
-require_once plugin_dir_path(__FILE__) . 'src/class-admin-settings.php';
+require_once plugin_dir_path(__FILE__) . 'src/admin-settings.php';
 require_once plugin_dir_path(__FILE__) . 'src/class-deactivator.php';
 require_once plugin_dir_path(__FILE__) . 'src/class-wa-integration.php';
 require_once plugin_dir_path(__FILE__) . 'src/helpers.php';
@@ -54,7 +54,7 @@ function wawp_enqueue_shortcode_css() {
 }
 
 // Create settings page
-$my_settings_page = new WAWP\Admin_Settings();
+$settings = new WAWP\Settings();
 
 // Create WA Integration instance
 $wa_integration_instance = new WAWP\WA_Integration();
